Tests/LibWeb/Text/expected/HTML/script-element-referrer-policy.txt

referrerPolicy = ''

referrerPolicy = 'no-referrer'

referrerPolicy = ''
